+++ packages/auth/client/components/AuthForm/index.js
@@ -1,31 +1,43 @@
 import React, { useState } from 'react'
 import { View } from 'react-native'
-import { observer } from 'startupjs'
+import { observer, emit, useLocal } from 'startupjs'
 import { H5, Content, Span } from '@startupjs/ui'
+import { SIGN_IN_URL, SIGN_UP_URL, RECOVER_PASS_URL } from '@startupjs/auth/isomorphic'
 import PropTypes from 'prop-types'
+import _get from 'lodash/get'
 import OrDivider from '../OrDivider'
-import { DEFAULT_FORMS_CAPTIONS, DEFAULT_FORMS_DESCRIPTIONS, FORM_COMPONENTS_KEYS, SIGN_IN_SLIDE, SIGN_UP_SLIDE } from '../../../isomorphic'
+import {
+  DEFAULT_FORMS_CAPTIONS,
+  DEFAULT_FORMS_DESCRIPTIONS,
+  FORM_COMPONENTS_KEYS,
+  SIGN_IN_SLIDE,
+  SIGN_UP_SLIDE,
+  RECOVER_PASSWORD_SLIDE
+} from '../../../isomorphic'
 import './index.styl'
 
 function AuthForm ({
-  captions,
-  descriptions,
+  configs,
   initSlide,
   socialButtons,
   localForms,
   hasRouting,
   onSuccess,
   onError,
-  onHandleError
+  onHandleError,
+  onChangeAuthPage
 }) {
+  const [search = ''] = useLocal('$render.search')
+
   const [activeSlide, setActiveSlide] = useState(initSlide)
   const LocalActiveForm = localForms ? localForms[FORM_COMPONENTS_KEYS[activeSlide]] : null
 
-  const _captions = Object.assign(DEFAULT_FORMS_CAPTIONS, captions)
-  const currentCaption = _captions[activeSlide]
+  // Config with titles, texts, custom components etc.
+  const config = _get(configs, activeSlide, {})
 
-  const _descriptions = Object.assign(DEFAULT_FORMS_DESCRIPTIONS, descriptions)
-  const currentDescription = _descriptions[activeSlide]
+  const caption = config.title || DEFAULT_FORMS_CAPTIONS[activeSlide]
+  const description = config.description || DEFAULT_FORMS_DESCRIPTIONS[activeSlide]
+  const localFormDescription = config.localFormDescription
 
   const renderSocialButtons = socialButtons.map((Component, index) => {
     return pug`
@@ -37,17 +49,45 @@ function AuthForm ({
   const needOrLine = renderSocialButtons && renderSocialButtons.length &&
     LocalActiveForm && [SIGN_IN_SLIDE, SIGN_UP_SLIDE].includes(activeSlide)
 
+  function _onChangeAuthPage (slide) {
+    // Catch case if we need update query params or do something else
+    if (onChangeAuthPage && hasRouting) {
+      onChangeAuthPage(slide)
+    } else {
+      if (hasRouting) {
+        switch (slide) {
+          case SIGN_IN_SLIDE:
+            emit('url', SIGN_IN_URL + search)
+            break
+          case SIGN_UP_SLIDE:
+            emit('url', SIGN_UP_URL + search)
+            break
+          case RECOVER_PASSWORD_SLIDE:
+            emit('url', RECOVER_PASS_URL + search)
+            break
+          default:
+            break
+        }
+      } else {
+        setActiveSlide(slide)
+      }
+    }
+  }
+
   return pug`
     Content
       if LocalActiveForm
-        if typeof currentCaption === 'string'
-          H5.caption= currentCaption
+        if typeof caption === 'string'
+          H5.caption= caption
         else
-          = currentCaption
-        if currentDescription
-          Span.description(variant='description')= currentDescription
+          = caption
 
-      if activeSlide !== 'recover'
+        if typeof description === 'string'
+          Span.description(variant='description')= description
+        else
+          = description
+
+      if [SIGN_IN_SLIDE, SIGN_UP_SLIDE].includes(activeSlide)
         View.buttons
           = renderSocialButtons
 
@@ -56,25 +96,29 @@ function AuthForm ({
 
       if LocalActiveForm
         View.form
+          if typeof localFormDescription === 'string'
+            Span.description(variant='description')= localFormDescription
+          else
+            = localFormDescription
           LocalActiveForm(
             onSuccess=onSuccess
             onError=onError
             onHandleError=onHandleError
-            onChangeAuthPage=hasRouting ? null : setActiveSlide
+            onChangeAuthPage=_onChangeAuthPage
           )
   `
 }
 
 AuthForm.propTypes = {
-  captions: PropTypes.object,
-  descriptions: PropTypes.object,
+  configs: PropTypes.object,
   initSlide: PropTypes.string,
   socialButtons: PropTypes.array,
   localForms: PropTypes.object,
   hasRouting: PropTypes.bool,
   onSuccess: PropTypes.func,
   onError: PropTypes.func,
-  onHandleError: PropTypes.func
+  onHandleError: PropTypes.func,
+  onChangeAuthPage: PropTypes.func
 }
 
 AuthForm.defaultProps = {