+++ packages/auth/client/helpers/finishAuth.js
@@ -3,11 +3,12 @@ import AsyncStorage from '@react-native-async-storage/async-storage'
 import RNRestart from 'react-native-restart'
 import { $root } from 'startupjs'
 
-export default async function finishAuth () {
-  const successRedirectUrl = $root.get('_session.auth.successRedirectUrl') || '/'
+export default async function finishAuth (redirectUrl) {
+  const successRedirectUrl = redirectUrl || $root.get('$render.query.redirectUrl') ||
+    $root.get('_session.auth.successRedirectUrl') || '/'
 
   if (Platform.OS === 'web') {
-    window.location.pathname = successRedirectUrl
+    window.location.href = successRedirectUrl
   } else {
     await AsyncStorage.setItem('successRedirectUrl', successRedirectUrl)
     RNRestart.Restart()