+++ packages/auth/client/helpers/onLogout.js
@@ -1,8 +1,9 @@
-import { NativeModules } from 'react-native'
+import { $root } from 'startupjs'
 import axios from 'axios'
 import { LOGOUT_URL } from '../../isomorphic'
 
-export default async function onLogout () {
+export default async function onLogout (redirectUrl) {
+  const _redirectUrl = $root.get('_session.auth.signInPageUrl') || redirectUrl || '/'
   await axios.get(LOGOUT_URL)
-  NativeModules.DevSettings.reload()
+  window.location.pathname = _redirectUrl
 }