+++ packages/ui/components/forms/Checkbox/checkbox.js
@@ -12,6 +12,7 @@ export default observer(function Checkbox ({
   style,
   className,
   value,
+  icon,
   disabled,
   onPress,
   ...props
@@ -51,7 +52,7 @@ export default observer(function Checkbox ({
     )
       Icon.checkbox-icon(
         styleName=[checkedStyleName]
-        icon=faCheck
+        icon= icon || faCheck
         size='s'
       )
       AnimatedView.checkbox-animation(