+++ styleguide/webpack.web.config.cjs
@@ -5,10 +5,11 @@ module.exports = getConfig(undefined, {
     'react-native-collapsible',
     '@startupjs/ui',
     '@startupjs/auth',
+    '@startupjs/auth-apple',
+    '@startupjs/auth-azuread',
     '@startupjs/auth-facebook',
     '@startupjs/auth-google',
     '@startupjs/auth-linkedin',
-    '@startupjs/auth-azuread',
     '@startupjs/auth-local'
   ],
   mode: 'react-native'